update department api and tests

+++ company_project/employee_manager/api/department.py
@@ -64,25 +64,32 @@ class DepartmentView(View):
         department_id = request_data.get('id')
         department_new_name = request_data.get('department_name')
 
-        try:
-            department = Department.objects.get(id=department_id)
-        except ObjectDoesNotExist:
-            status = 404
-            message = "Department not found"
+        if department_id and department_new_name:
+
+            try:
+                department = Department.objects.get(id=department_id)
+            except ObjectDoesNotExist:
+                status = 404
+                message = "Department not found"
+
+            else:
+                old_department_name = department.name
+                department.name = department_new_name
+                try:
+                    department.save()
+                except IntegrityError:
+                    status = 409
+                    message = "Department already exists"
+                else:
+                    status = 200
+                    message = "Department name {} changed to {}".format(
+                        old_department_name, department_new_name
+                    )
 
         else:
-            old_department_name = department.name
-            department.name = department_new_name
-            try:
-                department.save()
-            except IntegrityError:
-                status = 409
-                message = "Department already exists"
-            else:
-                status = 200
-                message = "Department name {} changed to {}".format(
-                    old_department_name, department_new_name
-                )
+            status = 400
+            message = 'Missing data'
+            response_data = {'content': message}
 
         response_data = {'content': message}
         return HttpResponse(
@@ -97,19 +104,26 @@ class DepartmentView(View):
         request_delete = json.loads(request.body)
         department_name = request_delete.get('department_name')
 
-        try:
-            department = Department.objects.get(name=department_name)
-        except ObjectDoesNotExist:
-            status = 404
-            message = "Department not found"
-        else:
-            if department.employee_set.count():  # if it has employees related
-                status = 400
-                message = "This department cannot be deleted: it has employees"
+        if department_name:
+            try:
+                department = Department.objects.get(name=department_name)
+            except ObjectDoesNotExist:
+                status = 404
+                message = "Department not found"
             else:
-                department.delete()
-                status = 200
-                message = "Department '{}' deleted".format(department_name)
+                # if it has employees related
+                if department.employee_set.count():
+                    status = 400
+                    message = "Department cannot be deleted: it has employees"
+                else:
+                    department.delete()
+                    status = 200
+                    message = "Department '{}' deleted".format(department_name)
+
+        else:
+            status = 400
+            message = 'Missing data'
+            response_data = {'content': message}
 
         response_data = {'content': message}
         return HttpResponse(
